Move active tile definitions to separate header
