+++ src/active_tile_data.h
@@ -4,11 +4,9 @@
 
 #include <string>
 #include "calendar.h"
-#include "point.h"
 
 class JsonObject;
 class JsonOut;
-class map;
 struct tripoint;
 class distribution_grid;
 
@@ -50,62 +48,6 @@ class active_tile_data
 
         virtual void store( JsonOut &jsout ) const = 0;
         virtual void load( JsonObject &jo ) = 0;
-};
-
-class solar_tile : public active_tile_data
-{
-    public:
-        /* In W */
-        int power;
-
-        void update_internal( time_point to, const tripoint &p, distribution_grid &grid ) override;
-        active_tile_data *clone() const override;
-        const std::string &get_type() const override;
-        void store( JsonOut &jsout ) const override;
-        void load( JsonObject &jo ) override;
-};
-
-class battery_tile : public active_tile_data
-{
-    public:
-        /* In J */
-        int stored;
-        int max_stored;
-
-        void update_internal( time_point to, const tripoint &p, distribution_grid &grid ) override;
-        active_tile_data *clone() const override;
-        const std::string &get_type() const override;
-        void store( JsonOut &jsout ) const override;
-        void load( JsonObject &jo ) override;
-
-        int get_resource() const;
-        int mod_resource( int amt );
-};
-
-class charger_tile : public active_tile_data
-{
-    public:
-        /* In W */
-        int power;
-
-        void update_internal( time_point to, const tripoint &p, distribution_grid &grid ) override;
-        active_tile_data *clone() const override;
-        const std::string &get_type() const override;
-        void store( JsonOut &jsout ) const override;
-        void load( JsonObject &jo ) override;
-};
-
-class vehicle_connector_tile : public active_tile_data
-{
-    public:
-        /* In absolute map square coordinates */
-        std::vector<tripoint> connected_vehicles;
-
-        void update_internal( time_point to, const tripoint &p, distribution_grid &grid ) override;
-        active_tile_data *clone() const override;
-        const std::string &get_type() const override;
-        void store( JsonOut &jsout ) const override;
-        void load( JsonObject &jo ) override;
 };
 
 // TODO: Better place for this
